requirements.txt
numpy
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pandas
scikit-learn
python-dotenv
pymysql
email-validator
fastapi-cors
pyjwt
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings==2.1.0